Analyze using a simplified limits of confidence interval design

In this simplified example of upper and lower confidence boundary designs, if it is likely that the treatment difference is above the Minimum Acceptable Value (MAV) then a Go decision is made. If a Go decision is not made, then if is is unlikely that the treatment difference is above the Target Value (TV) a No Go decision is made. In this example, the prop.test from base R is utilized to analyze the data and compute at user-specified confidence interval (dConfLevel). We set the defauly without user-specified variables to assume the MAV = 0.1 and TV=0.2. The team would like to make a Go decision if there is at least a 90 Using a frequentist CI an approximation to this design can be done by the logic described below. At an Interim Analysis, If the Lower Limit of the CI, denoted by LL, is greater than user-specified dLowerLimit then a Go decision is made. Specifically, if LL > UserParam$dLowerLimit --> Go If a Go decision is not made, then if the Upper Limit of the CI, denoted by UL, is less than user-specified dUpperLimit a No Go decision is made. Specifically, if UL < UserParam$dUpperLimit --> No Go Otherwise, continue to the next analysis. At the Final Analysis: If the Lower Limit of the CI, denoted by LL, is greater than dLowerLimit then a Go decision is made. Specifically, if LL > UserParam$dLowerLimit --> Go Otherwise, a No Go decision is made

Usage

AnalyzeUsingPropLimitsOfCI(SimData, DesignParam, LookInfo, UserParam = NULL)

Arguments

SimData

Data frame which consists of data generated in current simulation.

DesignParam

List of Design and Simulation Parameters required to perform analysis.

LookInfo

List containing Design and Simulation Parameters, which might be required to perform analysis.

UserParam

A list of user defined parameters in East. The default must be NULL. If UserParam is supplied, the list must contain the following named elements: UserParam$dLowerLimit - A value (0,1) that specifes the lower limit for the confidence interval. UserParam$dUpperLimit - A value (0,1) that specifies the upper limit for the confidence interval. UserParam$dConfLevel - A value (0,1) that specifies the confidence level for the prop.test function in base R.

Value

TestStat A double value of the computed test statistic

Decision An integer value: Decision = 0 --> No boundary crossed Decision = 1 --> Lower Efficacy Boundary Crossed Decision = 2 --> Upper Efficacy Boundary Crossed Decision = 3 --> Futility Boundary Crossed Decision = 4 --> Equivalence Boundary Crossed

ErrorCode An integer value: ErrorCode = 0 --> No Error

Note

In this example, the boundary information that is computed and sent from East is ignored in order to implement this decision approach.

Helpful Hints: There is often info that East sends to R that are not shown in a given example. It can be very helpful to save the input objects and then load them into your R session and inspect them. This can be done with the following R code in your function.

saveRDS( SimData, "SimData.Rds") saveRDS( DesignParam, "DesignParam.Rds" ) saveRDS( LookInfo, "LookInfo.Rds" )

The above code will save each of the input objects to a file so they may be examined within R.